Fixed bug '-d 10 option doesn't help frozen vikings on a strong machine'

+++ freeVikings/freevikings.rb
@@ -165,7 +165,7 @@ begin
         raise ArgumentError, "Argument of option --v-unit (-u) must be a real number higher than zero (given '#{argument}')"
       end
     when "--delay"
-      FreeVikings::OPTIONS['delay'] = argument.to_i / 1000
+      FreeVikings::OPTIONS['delay'] = argument.to_f / 1000
     when "--startpassword"
       unless FreeVikings.valid_location_password?(argument)
         raise "'#{argument}' is not a valid location password."
